Remove everything in ui/components/SignTransaction
This is the first step in eliminating all vestiges of the old
transaction signing flow. The feature flag can't be dropped yet, as
there are a few other entanglements left.

A few components were still in use:
- Several around Ledger status; these have been moved to SignerLedger/
  under the new signing structure.
- The loading screen for signature data was still in use; it's been
  moved to `Signing/SigningLoading`.
- The network account info top bar, which is static in the signing flow
  (instead of modifiable in the main wallet) was also in
  SignTransaction; it's been moved to `SigningNetworkAccountInfoTopBar`
  under `Signing`.
- The `SignTransactionDetailWarning` component, used to display warnings
  such as when assets are being transferred directly to a contract
  address; it's been moved to `TransactionSignatureDetailsWarning` under
  the `TransactionSignatureDetails` section of `Signing`.

Other than name adjustments, the moved components have no changes.
