+++ ui/components/Signing/Signer/SignerLedger/SignerLedgerConnect/ActivateBlindSigning.tsx
@@ -1,13 +1,13 @@
 import React, { ReactElement } from "react"
 import { useTranslation } from "react-i18next"
-import SignTransactionSlideUpContentLayout from "./SignTransactionSlideUpContentLayout"
+import SlideUpContentLayout from "./SlideUpLayout"
 
-export default function SignTransactionLedgerActivateBlindSigning(): ReactElement {
+export default function LedgerActivateBlindSigning(): ReactElement {
   const { t } = useTranslation("translation", {
     keyPrefix: "ledger.activation",
   })
   return (
-    <SignTransactionSlideUpContentLayout
+    <SlideUpContentLayout
       title={t("title")}
       helpMessage={t("helpMessage")}
       steps={[<>{t("step1")}</>, <>{t("step2")}</>, <>{t("step3")}</>]}
