+++ ui/components/Signing/Signer/SignerLedger/SignerLedgerConnect/LedgerNotConnected.tsx
@@ -1,13 +1,13 @@
 import React, { ReactElement } from "react"
 import { useTranslation } from "react-i18next"
-import SignTransactionSlideUpContentLayout from "./SignTransactionSlideUpContentLayout"
+import SlideUpContentLayout from "./SlideUpLayout"
 
-export default function SignTransactionMultipleLedgersConnected(): ReactElement {
+export default function LedgerNotConnected(): ReactElement {
   const { t } = useTranslation("translation", {
-    keyPrefix: "ledger.multipleLedgersConnected",
+    keyPrefix: "ledger.notConnected",
   })
   return (
-    <SignTransactionSlideUpContentLayout
+    <SlideUpContentLayout
       title={t("title")}
       helpMessage={t("helpMessage")}
       steps={[<>{t("step1")}</>, <>{t("step2")}</>, <>{t("step3")}</>]}
