+++ ui/components/Signing/Signer/SignerLedger/SignerLedgerConnect/index.tsx
@@ -1,11 +1,11 @@
 import { assertUnreachable } from "@tallyho/tally-background/lib/utils/type-guards"
 import React, { ReactElement } from "react"
-import SignTransactionLedgerActivateBlindSigning from "../../../../SignTransaction/SignTransactionLedgerActivateBlindSigning"
-import SignTransactionLedgerBusy from "../../../../SignTransaction/SignTransactionLedgerBusy"
-import SignTransactionLedgerNotConnected from "../../../../SignTransaction/SignTransactionLedgerNotConnected"
-import SignTransactionMultipleLedgersConnected from "../../../../SignTransaction/SignTransactionMultipleLedgersConnected"
-import SignTransactionWrongLedgerConnected from "../../../../SignTransaction/SignTransactionWrongLedgerConnected"
-import { SigningLedgerState } from "../../../../SignTransaction/useSigningLedgerState"
+import { SigningLedgerState } from "../../../../../hooks"
+import LedgerActivateBlindSigning from "./ActivateBlindSigning"
+import LedgerBusy from "./LedgerBusy"
+import LedgerNotConnected from "./LedgerNotConnected"
+import MultipleLedgersConnected from "./MultipleLedgersConnected"
+import WrongLedgerConnected from "./WrongLedgerConnected"
 
 type SignerLedgerConnectProps = {
   signingLedgerState: SigningLedgerState
@@ -16,23 +16,23 @@ export default function SignerLedgerConnect({
 }: SignerLedgerConnectProps): ReactElement {
   switch (signingLedgerState.state) {
     case "no-ledger-connected":
-      return <SignTransactionLedgerNotConnected />
+      return <LedgerNotConnected />
     case "wrong-ledger-connected":
       return (
-        <SignTransactionWrongLedgerConnected
+        <WrongLedgerConnected
           requiredAddress={signingLedgerState.requiredAddress}
         />
       )
     case "multiple-ledgers-connected":
-      return <SignTransactionMultipleLedgersConnected />
+      return <MultipleLedgersConnected />
     case "busy":
-      return <SignTransactionLedgerBusy />
+      return <LedgerBusy />
     case "available":
       // Note that if the ledger is available, this component expects to only
       // be rendered if arbitrary data signing is required AND disabled. We
       // don't guard with an additional check here, as the parent is expected
       // to verify for us.
-      return <SignTransactionLedgerActivateBlindSigning />
+      return <LedgerActivateBlindSigning />
     default:
       return assertUnreachable(signingLedgerState)
   }
